+++ presentation/src/main/java/co/orange/presentation/bank/BankActivity.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.setting.bank
+package co.orange.presentation.bank
 
 import android.os.Bundle
 import androidx.activity.viewModels
