+++ presentation/src/main/java/co/orange/presentation/delivery/AddressWebActivity.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.address
+package co.orange.presentation.delivery
 
 import android.content.Context
 import android.content.Intent
