+++ presentation/src/main/java/co/orange/presentation/delivery/DeliveryViewModel.kt
@@ -1,11 +1,11 @@
-package co.orange.presentation.setting.delivery
+package co.orange.presentation.delivery
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.orange.core.state.UiState
 import co.orange.domain.entity.response.AddressModel
 import co.orange.domain.repository.SettingRepository
-import co.orange.presentation.address.AddressActivity.Companion.DEFAULT_ID
+import co.orange.presentation.delivery.AddressActivity.Companion.DEFAULT_ID
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
