+++ presentation/src/main/java/co/orange/presentation/setting/AccountQuitDialog.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.setting.account
+package co.orange.presentation.setting
 
 import android.os.Bundle
 import android.view.View
